Pulled CharacterGroupGenerationPreferences out of GlyssenDblTextMetadata
Also moved Utilities/MessageBoxStrings to GlyssenEngine
