+++ Glyssen/Dialogs/CastSizePlanningViewModel.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using Glyssen.Bundle;
+using GlyssenEngine.Bundle;
 using GlyssenEngine.Character;
 using GlyssenEngine.VoiceActor;
