+++ Glyssen/Dialogs/ExportDlg.cs
@@ -5,8 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Windows.Forms;
 using Glyssen.Shared;
-using Glyssen.Utilities;
-using L10NSharp;
+using GlyssenEngine.Utilities;
 using L10NSharp.TMXUtils;
 using L10NSharp.UI;
 using SIL;
