+++ Glyssen/Dialogs/QuotationMarksDlg.cs
@@ -4,7 +4,6 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Windows.Forms;
-using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Controls;
 using Glyssen.Paratext;
@@ -12,8 +11,8 @@ using Glyssen.Properties;
 using Glyssen.Quote;
 using Glyssen.Shared;
 using Glyssen.Utilities;
+using GlyssenEngine.Bundle;
 using GlyssenEngine.Paratext;
-using GlyssenEngine.Utilities;
 using L10NSharp.TMXUtils;
 using L10NSharp.UI;
 using SIL;
