+++ Glyssen/Dialogs/ScriptureRangeSelectionDlg.cs
@@ -6,12 +6,11 @@ using System.Linq;
 using System.Windows.Forms;
 using System.Windows.Forms.VisualStyles;
 using DesktopAnalytics;
-using Glyssen.Bundle;
 using Glyssen.Controls;
 using Glyssen.Paratext;
 using Glyssen.Shared;
 using Glyssen.Utilities;
-using L10NSharp;
+using GlyssenEngine.Bundle;
 using L10NSharp.TMXUtils;
 using L10NSharp.UI;
 using SIL;
