+++ Glyssen/MainForm.cs
@@ -15,6 +15,7 @@ using Glyssen.Properties;
 using Glyssen.Rules;
 using Glyssen.Shared;
 using Glyssen.Utilities;
+using GlyssenEngine.Bundle;
 using GlyssenEngine.Paratext;
 using GlyssenEngine.Utilities;
 using L10NSharp;
