+++ Glyssen/ProjectBase.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using Glyssen.Shared;
 using Glyssen.Shared.Bundle;
-using L10NSharp;
 using SIL;
 using SIL.DblBundle;
 using SIL.Scripture;
