+++ Glyssen/Rules/Proximity.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using Glyssen.Bundle;
 using GlyssenEngine;
+using GlyssenEngine.Bundle;
 using GlyssenEngine.Character;
 using SIL.Scripture;
