+++ Glyssen/SampleProject.cs
@@ -5,6 +5,7 @@ using Glyssen.Bundle;
 using Glyssen.Quote;
 using Glyssen.Shared.Bundle;
 using GlyssenEngine;
+using GlyssenEngine.Bundle;
 using SIL.DblBundle.Text;
 using SIL.DblBundle.Usx;
 using SIL.WritingSystems;
