+++ GlyssenTests/Dialogs/CastSizeCalculationUtilities.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using Glyssen;
-using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Dialogs;
 using Glyssen.Rules;
+using GlyssenEngine.Bundle;
 using GlyssenEngine.Character;
 using NUnit.Framework;
 using SIL.Scripture;
